Two things note about this commit @mojodna
1) i did not make this work by overriding `as_json` on
the atlas, page, snapshot models because it seemed unnecessary
given that other parts of the code are requesting `as_json` to return
a different subset of json already:
https://github.com/fieldpapers/fp-web/blob/master/app/models/atlas.rb#L229
https://github.com/fieldpapers/fp-web/blob/master/app/models/atlas.rb#L265
https://github.com/fieldpapers/fp-web/blob/master/app/models/snapshot.rb#L212
https://github.com/fieldpapers/fp-web/blob/master/app/models/snapshot.rb#L194

2) this PR should not be accepted until we figure out why snapshots are not
quite working for `docker-compose.yml` dev environment. The `fp-tiler` does
not seem to create the `/snapshots/<slug>/index.json` on the last
step. still playing with this